refactor: consolidate SEC chunking functionality into vectorization.py
- Integrate section-aware chunking from sec_chunking_enhancer.py into vectorization.py
- Add SectionType enum and optimal chunk configurations for different SEC filing sections
- Add SECTION_AWARE chunking strategy and related functions
- Remove redundant sec_chunking_enhancer.py file (299 lines of duplicate code)
- Clean up references to the deleted file
- Maintain all existing functionality while eliminating code duplication
